ECER-1407: ReferenceTests separated from application tests, portal invitation test issue solved
